Add Magnet (not functioning)
Had to fix bridge again due to issue with tileset, had severe crash issues attempting this, added Magnet item and some other nodes and small collision to it.  May have to change items from static  body to kinematic in order to make magnet work with everything and the bat later.