Fixed Dragon Glitch, caused another
After fixing bitebox position dragon would never be able to reach it due to the node's position not changing with it, fixed this and now dragon will go to bait correctly instead of going past it and getting stuck.  The caused glitch is now while it is chasing the player, it will flip back and forth left and right if the player is directly above or below it, should be an easy fix to just make it face upwards correctly which is getting the y to read under 0.